+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+#### 0.3.0 (2019-03-24)
+
+##### New Features
+
+*  add default datetime format ([8fe8bd9a](https://github.com/lykmapipo/constants/commit/8fe8bd9adf2bd87e4d51d00d84e43bf6fcf3f3a7))
+*  add default time format ([7e697582](https://github.com/lykmapipo/constants/commit/7e69758289dd7474313c4268e0cfb4b60bac252f))
+*  add default date format ([1109e2f1](https://github.com/lykmapipo/constants/commit/1109e2f156179dab8077d9c57a5b71f269305a77))
+
+##### Refactors
+
+*  use mapToUpper to force uppercasing ([267e5595](https://github.com/lykmapipo/constants/commit/267e5595e5ea5d9639fc4f71f67415d373775234))
+
+##### Tests
+
+*  update default timezone assertions ([a4abfcc8](https://github.com/lykmapipo/constants/commit/a4abfcc82fe95c48060228434df0c92c4f70b33e))
+*  ensure default timezone test ([f798f6a2](https://github.com/lykmapipo/constants/commit/f798f6a2ee6118bd53cfd460891fbc793b759f92))
+
 #### 0.2.0 (2019-03-24)
 
 ##### Chores